refactor: rename trader-joe and vector-finance folders
