+++ .vscode/launch.json
@@ -310,7 +310,7 @@
         "tonpound",
         "tornado-cash",
         "tprotocol",
-        "traderjoe",
+        "trader-joe",
         "truefi",
         "trufin-protocol",
         "trustswap",
@@ -320,7 +320,7 @@
         "unsheth",
         "uwu-lend",
         "valas-finance",
-        "vector",
+        "vector-finance",
         "velodrome-v1",
         "velodrome-v2",
         "venus",
