+++ src/adapters/trader-joe/avalanche/index.ts
